+++ bolhadireta.h
@@ -1,6 +1,6 @@
 #ifndef BOLHADIRETA_H
 #define BOLHADIRETA_H
 
-void bubbleSort(int* vet, int tam, int* troca, int* comp);
+void bubbleSort(int* vet, int tam, long* troca, long* comp);
 
 #endif
