+++ radixsort.h
@@ -1,6 +1,6 @@
 #ifndef RADIXSORT_H
 #define RADIXSORT_H
 
-void radixSort(int* vet, int tam, int* comp, int* troca);
+void radixSort(int* vet, int tam, long* comp, long* troca);
 
 #endif
